+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.9.12] - 2026-02-15
+
+* 2024 edition, Rust 1.85
+
+* `SyncCell::as_ptr` is no longer unsafe, as you need
+  in any case unsafe code to dereference the pointer.
+
 ## [0.9.11] - 2024-11-27
 
 ### Improved